+++ relevanceai/visualise/projector.py
@@ -338,6 +338,9 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                 docs,
             )
 
+        if vector_label:
+            point_labels = self.get_field_across_documents(vector_label, docs)
+
         # Dimension reduce vectors
         vectors = np.array(self.get_field_across_documents(vector_field, docs))
         vectors_dr = DimReduction.dim_reduce(
@@ -391,6 +394,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             marker_size=marker_size,
             cluster=cluster,
             label_char_length=label_char_length,
+            vector_label=point_labels,
         )
 
         layout = self._generate_layout(plot_title=plot_title)
@@ -412,8 +416,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
 
         else:
             fig = go.Figure(data=plot_data, layout=layout)
-            fig.show()
-        return
+            return fig
 
     def _generate_plot_data(
         self,
@@ -429,6 +432,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             None,
         ],
         label_char_length: int,
+        vector_label: str,
     ):
         """ """
 
@@ -443,6 +447,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                         dims=dims,
                         marker_size=marker_size,
                         label_char_length=label_char_length,
+                        vector_label=vector_label,
                     )
                 )
 
@@ -455,6 +460,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                     dims=dims,
                     marker_size=marker_size,
                     label_char_length=label_char_length,
+                    vector_label=vector_label,
                 )
             )
 
@@ -468,6 +474,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             "backgroundcolor": "#ffffff",
             "showgrid": False,
             "showticklabels": False,
+            "showbackground": False,
         }
 
         axes_2d = {"title": "", "visible": False, "showticklabels": False}
@@ -491,7 +498,13 @@ class Projector(BatchAPIClient, _Base, DocUtils):
         return layout
 
     def _generate_plot_info(
-        self, embedding_df, hover_label, dims, marker_size, label_char_length
+        self,
+        embedding_df,
+        hover_label,
+        dims,
+        marker_size,
+        label_char_length,
+        vector_label,
     ):
         import plotly.graph_objects as go
 
@@ -506,10 +519,11 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             "x": embedding_df["x"],
             "y": embedding_df["y"],
             "showlegend": False,
-            "mode": "markers",
+            "mode": "markers+text",
             "marker": {"size": marker_size, "symbol": "circle", "opacity": 0.75},
             "customdata": custom_data,
             "hovertemplate": hovertemplate,
+            "text": vector_label,
         }
 
         if dims == 2: