+++ docsrc/source/changelog.rst
@@ -7,6 +7,8 @@ v0.26.6
 --------
 
 - Fix missing import in plotting since internalising plots
+- Add support for vector labels
+- Remove background axes from plot
 
 v0.26.5
 ---------